feat: sanitize error reporting in front end and make other minor changes
